py_server/requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn

asyncpg

pydantic
starlette
strawberry-graphql